Move App and Bot out of Chatbox
